refactoring libgdi
