+++ libgdi/decode.h
@@ -17,14 +17,15 @@
    limitations under the License.
 */
 
-#ifndef __GDI_DECODE_H
-#define __GDI_DECODE_H
+#ifndef __DECODE_H
+#define __DECODE_H
 
 #include <freerdp/freerdp.h>
-#include "rfx.h"
+
 #include "gdi.h"
+#include "rfx.h"
 
 void gdi_decode_frame(GDI *gdi, int x, int y, uint8 * data, uint32 length);
 void gdi_decode_data(GDI *gdi, uint8 * data, int data_size);
 
-#endif /* __GDI_DECODE_H */
+#endif /* __DECODE_H */
