+++ libgdi/gdi_32bpp.h
@@ -19,7 +19,6 @@
 
 #include <freerdp/freerdp.h>
 #include "gdi.h"
-#include "gdi_color.h"
 
 typedef void (*pSetPixel32_ROP2)(uint32 *pixel, uint32 *pen);
