+++ libgdi/gdi_bitmap.c
@@ -24,8 +24,7 @@
 #include <freerdp/freerdp.h>
 #include "gdi.h"
 
-#include "gdi_color.h"
-#include "gdi_window.h"
+#include "color.h"
 #include "gdi_32bpp.h"
 #include "gdi_16bpp.h"
 #include "gdi_8bpp.h"
