+++ libgdi/gdi_clipping.c
@@ -20,9 +20,10 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
 #include <freerdp/freerdp.h>
+
 #include "gdi.h"
+#include "gdi_region.h"
 
 #include "gdi_clipping.h"
