+++ libgdi/gdi_dc.c
@@ -22,9 +22,10 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
 #include <freerdp/freerdp.h>
+
 #include "gdi.h"
+#include "gdi_region.h"
 
 #include "gdi_dc.h"
