+++ libgdi/gdi_shape.c
@@ -20,13 +20,13 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
 #include <freerdp/freerdp.h>
-#include "gdi.h"
 
-#include "gdi_32bpp.h"
-#include "gdi_16bpp.h"
+#include "gdi.h"
 #include "gdi_8bpp.h"
+#include "gdi_16bpp.h"
+#include "gdi_32bpp.h"
+#include "gdi_bitmap.h"
 
 #include "gdi_shape.h"
